# docs/api_doc.yaml
openapi: 3.1.0
info:
  title: Digital Driving Log
  description: |-
    A comprehensive digital solution for managing shared vehicle driving logs and cost distribution.

    Some useful links:
    - [Digital-Driving-Log repository](https://github.com/Glueckhoch3/digital_driving_log)
    - [The source API definition for the DDL](https://github.com/Glueckhoch3/digital_driving_log/blob/main/docs/api_doc.yaml)

  license:
    name: License MIT
    url: https://github.com/Glueckhoch3/digital_driving_log/blob/main/LICENSE
  version: 0.0.2
servers:
  - url: https://localhost:8080/ddl/api

tags:
  - name: drives
    description: Access for logging and fetching driving logs.

  - name: costs
    description: Access for Cost data creation and retrieval.

paths:
  /drives:
    post:
      tags:
        - drives
      summary: Create a new drive
      description: Create a new drive entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDrive'
      responses:
        '201':
          description: Drive created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveDrive'
        '400':
          description: Invalid input

  /drives/{driveId}:
    get:
      tags:
        - drives
      summary: Get a drive by ID
      description: Retrieve a specific drive by its ID
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveDrive'
        '404':
          description: Drive not found
    put:
      tags:
        - drives
      summary: Update a drive
      description: Update an existing drive by ID
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: number
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutDrive'
      responses:
        '200':
          description: Drive updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveDrive'
        '404':
          description: Drive not found
    delete:
      tags:
        - drives
      summary: Delete a drive
      description: Delete a drive by ID
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '204':
          description: Drive deleted
        '404':
          description: Drive not found

  /vehicles/{carId}/drives:
    get:
      tags:
        - drives
      summary: Get all drives
      description: Retrieve a list of all drives
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveDrive'
        '404':
          description: No drives found

  /users/{userId}/drives:
    get:
      tags:
        - drives
      summary: Get all drives by user ID
      description: Retrieve all drives by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveDrive'
        '404':
          description: Drive not found

  /vehicles/{carId}/users/{userId}/drives:
    get:
      tags:
        - drives
      summary: Get all drives by user and vehicle ID
      description: Retrieve all drives by user and vehicle ID
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: number
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveDrive'
        '404':
          description: Drive not found

  /vehicles:
    get:
      tags:
        - vehicles
      summary: Get all vehicles
      description: Retrieve a list of all vehicles
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveCar'
        '404':
          description: No vehicles found
    post:
      tags:
        - vehicles
      summary: Create a new vehicle
      description: Create a new vehicle entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCar'
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveCar'
        '400':
          description: Invalid input

  /vehicles/{carId}:
    get:
      tags:
        - vehicles
      summary: Get a vehicle by ID
      description: Retrieve a specific vehicle by its ID
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveCar'
        '404':
          description: Car not found
    put:
      tags:
        - vehicles
      summary: Update a vehicle
      description: Update an existing vehicle by ID
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: number
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCar'
      responses:
        '200':
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveCar'
        '404':
          description: Vehicle not found
    delete:
      tags:
        - vehicles
      summary: Delete a vehicle
      description: Delete a vehicle by ID
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '204':
          description: Vehicle deleted
        '404':
          description: Vehicle not found

  /costs:
    get:
      tags:
        - costs
      summary: Get all costs
      description: Retrieve all costs
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveCost'
        '404':
          description: No costs found
    post:
      tags:
        - costs
      summary: Create a new cost
      description: Create a new cost entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCost'
      responses:
        '201':
          description: Cost created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveCost'
        '400':
          description: Invalid input


  /costs/{costId}:
    get:
      tags:
        - costs
      summary: Get a cost by ID
      description: Retrieve a specific cost by its ID
      parameters:
        - name: costId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveCost'
        '404':
          description: Cost not found
    put:
      tags:
        - costs
      summary: Update a cost
      description: Update an existing cost by ID
      parameters:
        - name: costId
          in: path
          required: true
          schema:
            type: number
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCost'
      responses:
        '200':
          description: Cost updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveCost'
        '404':
          description: Cost not found
    delete:
      tags:
        - costs
      summary: Delete a cost
      description: Delete a cost by ID
      parameters:
        - name: costId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '204':
          description: Cost deleted
        '404':
          description: Cost not found

  /vehicles/{carId}/costs:
    get:
      tags:
        - costs
      summary: Get all costs for a vehicle
      description: Retrieve all costs associated with a specific vehicle
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveCost'
        '404':
          description: Vehicle not found

  /users/{userId}/costs:
    get:
      tags:
        - costs
      summary: Get all costs for a user
      description: Retrieve all costs associated with a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveCost'
        '404':
          description: User not found

  /users:
    get:
      tags:
        - users
      summary: Get all users
      description: Retrieve a list of all users
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RetrieveUser'
        '404':
          description: No users found
    post:
      tags:
        - users
      summary: Create a new user
      description: Create a new user entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveUser'
        '400':
          description: Invalid input

  /users/{userId}:
    get:
      tags:
        - users
      summary: Get a user by ID
      description: Retrieve a specific user by their ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveUser'
        '404':
          description: User not found
    put:
      tags:
        - users
      summary: Update a user
      description: Update an existing user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutUser'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveUser'
        '404':
          description: User not found
    delete:
      tags:
        - users
      summary: Delete a user
      description: Delete a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
            format: int64
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found

components:
  schemas:
    CreateCar:
      type: object
      properties:
        name:
          type: string
        plateNumber:
          type: string
        ownerId:
          type: number
          format: int64
        brand:
          type: string
        data:
          type: string
      required:
         - name
         - plateNumber
         - ownerId

    RetrieveCar:
      type: object
      properties:
        carId:
          type: number
          format: int64
        name:
          type: string
        plateNumber:
          type: string
        ownerId:
          type: number
          format: int64
        brand:
          type: string
        data:
          type: string
      required:
         - carId
         - name
         - plateNumber
         - ownerId

    PutCar:
      type: object
      properties:
        name:
          type: string
        plateNumber:
          type: string
        ownerId:
          type: number
          format: int64
        brand:
          type: string
        data:
          type: string
      required:
         - name
         - plateNumber
         - ownerId

    CreateCost:
      type: object
      properties:
        carId:
          type: number
          format: int64
        buyerId:
          type: number
          format: int64
        transactionObject:
          type: string
        price:
          type: number
        dayOfTransaction:
          type: string
          format: date-time
        costType:
          type: string
        amount:
          type: number
        notes:
          type: string
      required:
         - carId
         - buyerId
         - transactionObject
         - price
         - dayOfTransaction
         - costType
         - amount

    RetrieveCost:
      type: object
      properties:
        costId:
          type: number
          format: int64
        carId:
          type: number
          format: int64
        buyerId:
          type: number
          format: int64
        transactionObject:
          type: string
        price:
          type: number
        dayOfTransaction:
          type: string
          format: date-time
        costType:
          type: string
        amount:
          type: number
        notes:
          type: string
      required:
         - costId
         - carId
         - buyerId
         - transactionObject
         - price
         - dayOfTransaction
         - costType
         - amount

    PutCost:
      type: object
      properties:
        carId:
          type: number
          format: int64
        buyerId:
          type: number
          format: int64
        transactionObject:
          type: string
        price:
          type: number
        dayOfTransaction:
          type: string
          format: date-time
        costType:
          type: string
        amount:
          type: number
        notes:
          type: string
      required:
         - carId
         - buyerId
         - transactionObject
         - price
         - dayOfTransaction
         - costType
         - amount

    CreateDrive:
      type: object
      properties:
        carId:
          type: number
          format: int64
        driveDate:
          type: string
          format: date-time
        distance:
          type: number
        driverId:
          type: number
          format: int64
        notes:
          type: string
      required:
        - carId
        - driveDate
        - distance
        - driverId

    RetrieveDrive:
      type: object
      properties:
        driveId:
          type: number
          format: int64
        carId:
          type: number
          format: int64
        driveDate:
          type: string
          format: date-time
        distance:
          type: number
        driverId:
          type: number
          format: int64
        notes:
          type: string
      required:
        - driveId
        - carId
        - driveDate
        - distance
        - driverId

    PutDrive:
      type: object
      properties:
        carId:
          type: number
          format: int64
        driveDate:
          type: string
          format: date-time
        distance:
          type: number
        driverId:
          type: number
          format: int64
        notes:
          type: string
      required:
        - carId
        - driveDate
        - distance
        - driverId

    CreateUser:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        driverLicense:
          type: boolean
        birthday:
          type: string
          format: date
      required:
        - firstname
        - lastname
        - driverLicense

    RetrieveUser:
      type: object
      properties:
        userId:
          type: number
          format: int64
        firstname:
          type: string
        lastname:
          type: string
        driverLicense:
          type: boolean
        birthday:
          type: string
          format: date
      required:
         - userId
         - firstname
         - lastname
         - driverLicense

    PutUser:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        driverLicense:
          type: boolean
        birthday:
          type: string
          format: date
      required:
        - firstname
        - lastname
        - driverLicense

    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message